DrumBot/requirements.txt
opencv-python>=4.9.0
mediapipe>=0.10.14
pyserial>=3.5
mido>=1.3.2
python-rtmidi>=1.5.8
numpy>=1.24
pytest>=7.4
